+++ Tests/TransferToArchiveTests.swift
@@ -1,0 +1,18 @@
+//
+//  TransferToArchiveTests.swift
+//
+// Copyright 2022, 2023  OpenAlloc LLC
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+//
+
+import CoreData
+
+@testable import GroutLib
+import XCTest
+
+final class TransferToArchiveTests: TestBase {
+    func testRoutineKeepAt() throws {}
+}
